backend/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
DateTime==6.0
distro==1.9.0
dj-database-url==3.1.2
Django==6.0.3
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
groq==1.0.0
gunicorn==25.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jmespath==1.1.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==1.2.17
langgraph==1.0.10
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.9
langsmith==0.7.13
orjson==3.11.7
opencv-python-headless==4.10.0.84
ormsgpack==1.12.2
packaging==26.0
pillow==12.1.1
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.12.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
tenacity==9.1.4
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.41.0
whitenoise==6.12.0
xxhash==3.6.0
zope.interface==8.2
zstandard==0.25.0
